requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
gunicorn==21.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
SQLAlchemy==2.0.29
google-generativeai==0.8.5
python-dotenv==1.1.1 